backend/pyproject.toml
[project]
name = "signal-backend"
version = "0.1.0"
description = "Signal backend — FastAPI + FastMCP + Cosmos DB (single process on VPS)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "fastmcp>=2.6.0",
    "azure-cosmos>=4.7.0",
    "boto3>=1.34.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
package = false
